requirements.txt
redis==3.2.1
python-dotenv==0.10.3
python-telegram-bot==11.1.0
vk-api==11.5.0
